# api/openapi.yaml
openapi: 3.0.3
info:
  title: Zyad Cloud API
  version: 0.1.0
paths:
...
  /api/v1/admin/landing-pages:
    get:
      summary: List landing pages
      operationId: listLandingPages
      tags:
        - Landing Pages
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Landing pages retrieved successfully
          content:
            application/json:
              schema:
                type: object
    post:
      summary: Create landing page
      operationId: createLandingPage
      tags:
        - Landing Pages
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Landing page created successfully
          content:
            application/json:
              schema:
                type: object
  /api/v1/admin/landing-pages/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get landing page detail
      operationId: getLandingPage
      tags:
        - Landing Pages
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Landing page retrieved successfully
          content:
            application/json:
              schema:
                type: object
    patch:
      summary: Update landing page
      operationId: updateLandingPage
      tags:
        - Landing Pages
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Landing page updated successfully
          content:
            application/json:
              schema:
                type: object
    delete:
      summary: Delete landing page
      operationId: deleteLandingPage
      tags:
        - Landing Pages
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Landing page deleted successfully
          content:
            application/json:
              schema:
                type: object
  /api/v1/admin/landing-pages/{id}/access:
    put:
      summary: Update landing page access
      operationId: updateLandingPageAccess
      tags:
        - Landing Pages
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Landing page access updated successfully
  /api/v1/admin/landing-pages/{id}/duplicate:
    post:
      summary: Duplicate landing page
      operationId: duplicateLandingPage
      tags:
        - Landing Pages
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Landing page duplicated successfully
  /api/v1/admin/landing-pages/{id}/archive:
    post:
      summary: Archive landing page
      operationId: archiveLandingPage
      tags:
        - Landing Pages
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Landing page archived successfully
  /api/v1/admin/landing-pages/{id}/restore:
    post:
      summary: Restore landing page
      operationId: restoreLandingPage
      tags:
        - Landing Pages
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Landing page restored successfully
  /api/v1/admin/landing-pages/{id}/seo:
    patch:
      summary: Update landing page SEO
      operationId: updateLandingPageSeo
      tags:
        - Landing Pages
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Landing page SEO updated successfully
  /api/v1/admin/landing-pages/{id}/publish:
    post:
      summary: Publish landing page
      operationId: publishLandingPage
      tags:
        - Landing Pages
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Landing page published successfully
  /api/v1/admin/landing-pages/{id}/unpublish:
    post:
      summary: Unpublish landing page
      operationId: unpublishLandingPage
      tags:
        - Landing Pages
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Landing page unpublished successfully
  /api/v1/admin/landing-pages/{id}/schedule:
    put:
      summary: Schedule landing page publish or unpublish
      operationId: scheduleLandingPage
      tags:
        - Landing Pages
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Landing page schedule updated successfully
    delete:
      summary: Cancel landing page pending schedules
      operationId: cancelLandingPageSchedule
      tags:
        - Landing Pages
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Landing page schedule cancelled successfully
...
  /api/v1/admin/landing-pages/{id}/revisions:
    get:
      summary: List landing page revisions
      operationId: listLandingPageRevisions
      tags:
        - Landing Revisions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Landing page revisions retrieved successfully
...
  /api/v1/admin/landing-pages/{id}/revisions/{revision}/restore:
    post:
      summary: Restore landing page revision
      operationId: restoreLandingPageRevision
      tags:
        - Landing Revisions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: revision
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Landing page revision restored successfully
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...